@BastiaanOlij 2023-07-18
Couple small things (already mentioned in chat but repeating here so it doesn't get lost).

We should remove `is_xr_enabled` from the system, thought that probably means we need to port this solution to the mobile renderer as well. `is_xr_enabled` was our old way of creating a multiview shader group, it's actually linked to a setting called "enable xr shaders".

Instead we can simply check for `viewcount > 1` at the start of `_render_scene` and if set enable our multiview group.

My other problem, and that might be a little harder to fix, is that if we have a single viewport, `viewcount > 1` is true and we're using advanced features, all shader groups will be enabled, while the only shader group that may be used is `SHADER_GROUP_ADVANCED_MULTIVIEW`.