@clayjohn 2023-07-21
> Couple small things (already mentioned in chat but repeating here so it doesn't get lost).
> 
> We should remove `is_xr_enabled` from the system, thought that probably means we need to port this solution to the mobile renderer as well. `is_xr_enabled` was our old way of creating a multiview shader group, it's actually linked to a setting called "enable xr shaders".
> 
> Instead we can simply check for `viewcount > 1` at the start of `_render_scene` and if set enable our multiview group.
> 
> My other problem, and that might be a little harder to fix, is that if we have a single viewport, `viewcount > 1` is true and we're using advanced features, all shader groups will be enabled, while the only shader group that may be used is `SHADER_GROUP_ADVANCED_MULTIVIEW`.

Pushed some changes to substantially address this. 

1) Removing ``is_xr_enabled`` is way outside the scope of this PR. To remove it, we need to update the shader_compiler to detect whether a given shader is being used in XR and then compile the XR-supporting variant. Right now we use ``is_xr_enabled`` for that, but we need to find something better.
2) I added code to enable the MULTIVIEW versions when ``view_count > 1`` regardless of ``is_xr_enabled``. However, I left the original code that checks ``is_xr_enabled`` as compiling up front will be preferable
3) I added an optional argument to ``enable_advanced_shader`` so that only ``SHADER_GROUP_ADVANCED_MULTIVIEW`` is enabled if using ``viewcount > 1``. ``SHADER_GROUP_ADVANCED`` will be enabled as soon as a non-XR viewport needs it.

In my opinion, this change should be ready to merge once it passes CI