@BastiaanOlij 2023-07-31
@clayjohn looks good to me, I agree that retiring `is_xr_enabled` is maybe too much to ask for this PR as that would also result in requiring similar changes in the compatibility renderer. Maybe we should add a comment inside of the method that we are looking at retiring it?

I think it also makes sense if we have project settings to pre-enable each shader group. If the user knows a shader group should be included always and is willing to have the cost up front,