@AThousandShips 2023-08-01
This looks like it causes a lot of `Attempted to free invalid ID: -9222835861727476870` errors on closedown