@hudmarc 2023-08-17
Could this help with super slow shader compile times on html5 builds with ANGLE? see https://github.com/godotengine/godot/issues/72584