@akien-mga 2023-08-17
No that's unrelated, this is for the RenderingDevice backends (Vulkan, D3D 12), not OpenGL/WebGL.